requirements.txt
altgraph==0.16.1
atomicwrites==1.2.1
attrs==18.2.0
colorama==0.4.1
coverage==4.5.2
future==0.17.1
keyboard==0.13.2
macholib==1.11
more-itertools==5.0.0
numpy==1.15.4
opencv-python==3.4.5.20
pathlib==1.0.1
pefile==2018.8.8
Pillow==5.4.0
pluggy==0.8.1
py==1.7.0
PyInstaller==3.4
pynput==1.4
pytest==4.1.1
pytest-cov==2.6.1
pywin32-ctypes==0.2.0
screeninfo==0.3
six==1.12.0
testfixtures==6.4.3
